requirements.txt
streamlit
yfinance
pandas
plotly
PyGithub
bcrypt
